requirements.txt
Flask
PyMuPDF
gunicorn
zipstream-ng==1.6.0
